Fix segmentation fault by improving tree balancing
Previously, creating a tree with for example 1000 identical points would lead
to a segmentation fault, because the balancing on the average would place all
values on one side. Now, the code checks for such bad balancing, and in that
case falls back to the 'perfect' balancing scheme that was present but
disabled.

This approach keeps the same performance as before (balancing on average seems
to be a bit faster), but avoids the segmentation fault.